Landing page links to chapters
Optimally when you click on docs - you immediately know where to look to find your answer. I think adding these links to the top would improve the landing page, because you don't have to scroll to find the user guide, for instance. Ideally the screenshot would be to the right of the link, or the links would be buttons spanning horisontally, because in my proposal I think the page looks imbalanced with narrow lists on top and the huge image below (this is on a wide screen device). As it is now, I think it looks better on mobile, the image is smaller.