Correct Rate Limit reset documentation
I've been working on a project using the Rocket Chat API and have noticed a couple of places where the documentation is incorrect or out of date. First, the three rate limit properties seem to be sent with _every_ API call (tested on `/api/v1/channels.history` at least) and second the rate limit reset header seems to be provided in _milli_seconds, not seconds. Example from a server I'm on: `1562287945706`.

Are these values configurable by the server owner (in which case I assume additional doc changes are necessary)?