Add missing install command in instructions
I found this problem while trying to run the command for installing the required tools via WSL 2 shell.

The 'install' keyword is missing, which causes the installation process to fail. This minor text edit includes this keyword so that it doesn't block users during the installation process while they follow along with the guide.