Revised the database dump cp command for backup (#2019)
Assuming the data dump volume configuration is `- ./data/dump:/dump` on Docker Compose, the host side location would be `/opt/rocketchat/data/dump/` since a `/data` directory is non-existent on Linux. It could also be `cp data/dump/mmap ~/mongo-mmap-dump.gzip` but for the purpose of the complete doc, the proposed change makes it more clear.